refactor(app): :recycle: Added project support for golangci-lint: cod…

+++ dms3client/client_config.go
@@ -18,5 +18,5 @@ type structSettings struct {
 type structServer struct {
 	IP            string
 	Port          int
-	CheckInterval int
+	CheckInterval uint16
 }
